move ColumnKwargs and fix digits argument
